% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armaStats.R
\name{armaTrueacf}
\alias{armaTrueacf}
\title{Statistics of the True ARMA Process}
\usage{
armaRoots(coefficients, n.plot = 400, digits = 4, ...)
armaTrueacf(model, lag.max = 20, type = c("correlation", "partial", "both"),
            doplot = TRUE)
}
\arguments{
\item{model}{\link{armaTrueacf} -
a specification of the ARMA model with two elements: \code{model$ar} is the
vector of the AR coefficients, and \code{model$ma} is the vector of the MA
coefficients.}

\item{lag.max}{\link{armaTrueacf} -
maximum number of lags at which to calculate the acf or pacf, an
integer value by default 20.}

\item{type}{\link{armaTrueacf} -
a character string, "correlation" to compute the true autocorrelation
function, "partial" to compute the true partial autocorrelation function,
or "both" if both functions are desired. The start of one of the strings
will suffice.}

\item{doplot}{\link{armaRoots} -
a logical. Should a plot be displayed?}

\item{coefficients}{\link{armaRoots} -
a numeric vector with the coefficients of the characterisitic polynomial.}

\item{digits}{\link{armaRoots} -
output precision, an integer value.}

\item{n}{\link{armaSim} -
an integer value setting the length of the series to be simulated
(optional if \code{innov} is provided). The default value is 100.}

\item{n.plot}{\link{armaRoots} -
the number of data points to plot the unit circle; an integer value.}

\item{...}{additional arguments to be passed.}
}
\value{
\code{armaRoots} and returns a three column data frame with the real,
the imaginary part and the radius of the roots.
The number of rows corresponds to the coefficients.

\code{armaTrueacf} returns a two column data frame with the lag and
the correlation function.
}
\description{
A collection and description of functions to compute statistics of
a true ARMA time series process.

The Functions are:

\code{armaRoots}	Roots of the characteristic ARMA polynomial.
\code{armaTrueacf}	True autocorrelation function of an ARMA process.
}
\examples{

# armaRoots - Calculate and plot the roots of an ARMA process:
par(mfrow = c(2, 2), cex = 0.7)
coefficients = c(-0.5, 0.9, -0.1, -0.5)
armaRoots(coefficients)

# armaTrueacf
model = list(ar = c(0.3, +0.3), ma = 0.1)
armaTrueacf(model)
model = list(ar = c(0.3, -0.3), ma = 0.1)
armaTrueacf(model)
}
