% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSeries.R
\name{.modelSeries}
\alias{.modelSeries}
\title{.modelSeries}
\usage{
.modelSeries(formula, data, fake = FALSE, lhs = FALSE)
}
\arguments{
\item{formula}{some formula}

\item{data}{a timeSeries, a data.frame or a numeric vector.}

\item{fake}{default is \code{FALSE}}

\item{lhs}{default is \code{FALSE}}
}
\value{
no return value.
}
\description{
Models a time Series object to use formulas
}
\details{
Time Series Modelling
Regression Modelling
Data Management
}
