% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgnStats.R
\name{fgnTrueacf}
\alias{fgnTrueacf}
\title{Statistics of the True LDR Process}
\usage{
fgnTrueacf(n = 100, H = 0.7)
fgnTruefft(n = 100, H = 0.7)
fgnStatsSlider()

farimaTrueacf(n = 100, H = 0.7)
farimaTruefft(n = 100, H = 0.7)farimaTruefft(n = 100, H = 0.7)
farimaStatsSlider()
}
\arguments{
\item{n}{number of data points to be simulated, a numeric value, by default 100.}

\item{H}{the Hurst exponent, a numeric value between 0.5 and 1, by default 0.7.}
}
\value{
\code{fgnTrueacf} and \code{farimaTrueacf} return the true covariance of
an FGN or FARIMA time series process.

\code{fgnTruefft} and \code{farimaTruefft} return the true spectrum of an FGN
or FARIMA time series process.
}
\description{
A collection and description of functions to investigate the true statistics of the long
range dependence or long memory behavior of an univariate time series process.

The Functions are:

Functions to model the true autocorrelation function and spectrum:

\code{fgnTrueacf}	Returns true FGN covariances.
\code{fgnTruefft}	returns true FGN fast Fourier transform.
\code{fgnStatsSlider}	returns a plot of true FGN Statistics.
\code{farimaTrueacf}	returns true FARIMA covariances.
\code{farimaTruefft}	returns true FARIMA fast Fourier transform.
\code{farimaStatsSlider}	returns a plot of true FARIMA Statistics.
}
\details{
Functions to Model True Correlations and Spectrum:

The functions \code{fgnTrueacf} and \code{farimaTrueacf} return the true covariances of
an FGN and Gaussian FARIMA(0,d,0) time series process.

The functions \code{fgnTruefft} and \code{farimaTruefft} return the true fast Fourier
transform of an FGN and Gaussian FARIMA(0,d,0) time series process.
}
\examples{

# fgnTrueacf
fgnTrueacf(n = 20, H = 0.8)

# fgnTruefft
fgnTruefft(n = 20, H = 0.8)

# farimaTrueacf
farimaTrueacf(n = 20, H = 0.8)

# farimaTruefft
farimaTruefft(n = 20, H = 0.8)
}
